game-theory/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.8
stripe==7.8.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# GTO Strategy Guidance Dependencies
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
requests==2.31.0
ccxt==4.1.58
web3==6.12.0
websockets==12.0
redis==5.0.1
aioredis==2.0.1